Hardware Aware Mamba + Long Term Forecasting
