+++ exp/exp_basic.py
@@ -2,7 +2,7 @@ import os
 import torch
 from models import Autoformer, Transformer, TimesNet, Nonstationary_Transformer, DLinear, FEDformer, \
     Informer, LightTS, Reformer, ETSformer, Pyraformer, PatchTST, MICN, Crossformer, FiLM, iTransformer, \
-    Koopa, TiDE, FreTS, Mamba
+    Koopa, TiDE, FreTS, MambaSimple, Mamba
 
 
 class Exp_Basic(object):
@@ -28,6 +28,7 @@ class Exp_Basic(object):
             'Koopa': Koopa,
             'TiDE': TiDE,
             'FreTS': FreTS,
+            'MambaSimple': MambaSimple,
             'Mamba': Mamba,
         }
         self.device = self._acquire_device()
