+++ scripts/long_term_forecast/Mamba_all.sh
@@ -1,0 +1,4 @@
+./scripts/long_term_forecast/ECL_script/Mamba.sh | tee mamba_all.txt
+./scripts/long_term_forecast/Traffic_script/Mamba.sh | tee mamba_all.txt -a
+./scripts/long_term_forecast/Exchange_script/Mamba.sh | tee mamba_all.txt -a
+./scripts/long_term_forecast/Weather_script/Mamba.sh | tee mamba_all.txt -a
